common : extract trie/ac to a separate file
